requirements.txt
wheel
fastapi
uvicorn
python-dotenv
gtts
pysoundfile
pydantic
simpleaudio
pydub
ffmpeg
pyautogui
pyscreeze
Pillow
python-multipart
PyGithub
requests
pytest
httpx
pytest-mock